[Windows] Make contracts branch build on MSVC/LLP64
EricWF's contracts-nightly branch is developed on Linux and relies on
LP64 and POSIX assumptions that break the Windows (clang-cl) build:

- Scope::ScopeFlags defines ContractAssertScope = 0x100000000 (bit 32),
  which only fits a 64-bit `unsigned long` under LP64. Under LLP64
  (Windows) `unsigned long` is 32-bit, so the flag was truncated to 0 and
  the scope-flag plumbing narrowed it away. Widen ScopeFlags and its
  plumbing (Scope, Parser, ParseStmt) to `unsigned long long`, which is
  64-bit on both data models.
- EricWFDebug.cpp includes <sys/wait.h> and uses fork()/wait(); guard the
  POSIX-only fork-based debugging aid behind !_WIN32.

+++ clang/include/clang/Parse/Parser.h
@@ -403,7 +403,7 @@ public:
     // ParseScope - Construct a new object to manage a scope in the
     // parser Self where the new Scope is created with the flags
     // ScopeFlags, but only when we aren't about to enter a compound statement.
-    ParseScope(Parser *Self, unsigned long ScopeFlags, bool EnteredScope = true,
+    ParseScope(Parser *Self, unsigned long long ScopeFlags, bool EnteredScope = true,
                bool BeforeCompoundStmt = false)
         : Self(Self) {
       if (EnteredScope && !BeforeCompoundStmt)
@@ -438,7 +438,7 @@ public:
 
   public:
     MultiParseScope(Parser &Self) : Self(Self) {}
-    void Enter(unsigned long ScopeFlags) {
+    void Enter(unsigned long long ScopeFlags) {
       Self.EnterScope(ScopeFlags);
       ++NumScopes;
     }
@@ -452,7 +452,7 @@ public:
   };
 
   /// EnterScope - Start a new scope.
-  void EnterScope(unsigned long ScopeFlags);
+  void EnterScope(unsigned long long ScopeFlags);
 
   /// ExitScope - Pop a scope off the scope stack.
   void ExitScope();
@@ -839,14 +839,14 @@ private:
   /// RAII object used to modify the scope flags for the current scope.
   class ParseScopeFlags {
     Scope *CurScope;
-    unsigned long OldFlags = 0;
+    unsigned long long OldFlags = 0;
     ParseScopeFlags(const ParseScopeFlags &) = delete;
     void operator=(const ParseScopeFlags &) = delete;
 
   public:
     /// Set the flags for the current scope to ScopeFlags. If ManageFlags is
     /// false, this object does nothing.
-    ParseScopeFlags(Parser *Self, unsigned long ScopeFlags, bool ManageFlags = true);
+    ParseScopeFlags(Parser *Self, unsigned long long ScopeFlags, bool ManageFlags = true);
 
     /// Restore the flags for the current scope to what they were before this
     /// object overrode them.
@@ -7362,7 +7362,7 @@ public:
   /// [GNU]   '__label__' identifier-list ';'
   /// \endverbatim
   ///
-  StmtResult ParseCompoundStatement(bool isStmtExpr, unsigned long ScopeFlags);
+  StmtResult ParseCompoundStatement(bool isStmtExpr, unsigned long long ScopeFlags);
 
   /// Parse any pragmas at the start of the compound expression. We handle these
   /// separately since some pragmas (FP_CONTRACT) must appear before any C
